+++ src/pages/home/feedHashtag.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import styled from "styled-components";
 import Hashtag from "components/hashtag";
+import { v4 as uuid } from "uuid";
 
 function FeedHashtag({ hashtag = [] }) {
   return (
     <StTagBox>
       {hashtag.map((tag) => {
-        return <Hashtag hashtag={true} content={tag} size={"sm"} onClick={() => alert("안녕")} />;
+        return <Hashtag hashtag={true} content={tag} size={"sm"} key={uuid()} />;
       })}
     </StTagBox>
   );
